Refactor project structure into modules
