+++ index.html
@@ -30,6 +30,6 @@
     <button id="clear-all">Eliminar todas</button>
     <ul id="flashcard-list"></ul>
     </div>
-    <script src="script.js"></script>
+    <script type="module" src="src/app.js"></script>
 </body>
 </html>
