+++ package.json
@@ -2,7 +2,7 @@
   "name": "proyecto_v1",
   "version": "1.0.0",
   "description": "Esta es una aplicación web sencilla para gestionar flashcards sin utilizar frameworks. Permite crear dos tipos de tarjetas:",
-  "main": "script.js",
+  "main": "src/app.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "serve ."
